Fixed all warnings

+++ src/Time/Time.cpp
@@ -2,15 +2,15 @@
 
 const double Time::MillisecondToSeconds = 1e-9;
 
-double Time::deltaTime = 0;
+float Time::deltaTime = 0;
 unsigned int Time::frameCount = 0;
 nanoseconds Time::initTime = {};
 
-double Time::GetTimeSinceStart()
+float Time::GetTimeSinceStart()
 {
 	auto now = clock.now();
 	auto sinceEpoch = now.time_since_epoch();
 	long long millisecondsSinceClockStart = (sinceEpoch - Time::initTime).count();
 
-	return millisecondsSinceClockStart * MillisecondToSeconds;
+	return float(millisecondsSinceClockStart * MillisecondToSeconds);
 }
