+++ src/Time/Time.h
@@ -12,15 +12,15 @@ class Time
 {
 public:
 	inline static unsigned int GetFrameCount() { return frameCount; }
-	inline static double GetDeltaTime() { return deltaTime; }
+	inline static float GetDeltaTime() { return deltaTime; }
 
-	static double GetTimeSinceStart();
+	static float GetTimeSinceStart();
 	
 private:
 	static unsigned int frameCount;
 	static high_resolution_clock clock;
 	static nanoseconds initTime;
-	static double deltaTime;
+	static float deltaTime;
 
 	static const double MillisecondToSeconds;
